requirements.txt
asgiref==3.7.2
attrs==23.1.0
autoflake==2.2.1
backports.zoneinfo==0.2.1
black==23.12.1
click==8.1.7
coverage==7.4.0
Django==4.2.8
django-crontab==0.7.1
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
typing-extensions==4.9.0
uritemplate==4.1.1
zipp==3.17.0
